--- runproj.go
@@ -1,10 +1,0 @@
-package main
-
-func main() {
-	StartAPI()
-	// args := os.Args[1:]
-
-	// sets := GetConfigContent()
-
-	// ExecuteSelectedSets(sets, args)
-}
